block and stream serialisation of interned values:
A new serialisation type is introduced for interned values.

For every value to be serialised we check whether it is an interned
value. If so we serialise the associated symbol followed by the value.

On deserialisation we deserialise the symbol, create an empty value
and intern it, just as if the |intern| primitive was called, obtaining
whatever value results from that. We then deserialize into that value.

Note that this means block and stream serialisation will replace the
*contents* of existing interned values.