* fixed |load| error locations to point to beginning of read expression We extract the annotations from expression we'v just read, and if they are absent we fall back to the input port's location before the expression was read.
This mostly fixes bug 1458196 but is still not perfect. Specifically,
if we hit the fall-back case - e.g. because annotations are turned off
or the expression is not a pair - then there can be white space and
comments between the reported location and the start of the actualy
expression that caused the error. Fixing that requires major surgery
though because the handling of multi-line comments is deeply entangled
with expression parsing.