* Fixed fatal error in de-optimisation of letrec. See bug 1509504. LetrecExp.eval builds a new r.lcl from the existing r.lcl and other information, and then delegates to AppExp.eval. De-optimisation caused the entire process to be repeated. Since the interpreter state had already been modified by the earlier call this resulted in incorrect and sometimes fatal behaviour.
The fix is to only repeat the execution of what is strictly necessary.

The change for this - replacing re-execution via the interpreter loop
by an explicit loop - is really only needed on AppExp. However, for
consistency I've also applied it to the other Exps. The change does
not appear to impede performance.